requirements.txt
# === Test Frameworks ===
pytest==8.4.2
pytest-html==4.1.1
pytest-selenium==4.1.0

# === Browser Automation ===
selenium==4.25.0


# === Document and Excel Handling ===
python-docx==1.2.0
openpyxl==3.1.5

